requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
sqlparse==0.4.3
#uWSGI==2.0.20
